+++ src/components/ProductList/index.js
@@ -1,1 +1,1 @@
-export { default } from "./ProductList";
+export { default } from "./ProductListContainer";
